+++ src/js/page/odoo/commands/common/__all__.mjs
@@ -45,6 +45,7 @@ import cmdWS from './ws';
 import cmdURL from './url';
 import cmdInfo from './info';
 import cmdNotify from './notify';
+import cmdExportFile from './exportfile';
 import type VMachine from '@trash/vmachine';
 
 export default function (vm: VMachine) {
@@ -91,4 +92,5 @@ export default function (vm: VMachine) {
   vm.registerCommand('url', cmdURL());
   vm.registerCommand('info', cmdInfo());
   vm.registerCommand('notify', cmdNotify());
+  vm.registerCommand('exportfile', cmdExportFile());
 }