+++ src/js/page/terminal/commands/__all__.mjs
@@ -6,7 +6,6 @@ import cmdChrono from './chrono';
 import cmdClear from './clear';
 import cmdContextTerm from './context_term';
 import cmdDis from './dis';
-import cmdExportFile from './exportfile';
 import cmdExportVar from './exportvar';
 import cmdGenFile from './genfile';
 import cmdHelp from './help';
@@ -26,7 +25,6 @@ export default function (vm: VMachine) {
   vm.registerCommand('context_term', cmdContextTerm());
   vm.registerCommand('alias', cmdAlias());
   vm.registerCommand('exportvar', cmdExportVar());
-  vm.registerCommand('exportfile', cmdExportFile());
   vm.registerCommand('chrono', cmdChrono());
   vm.registerCommand('jobs', cmdJobs());
   vm.registerCommand('toggle_term', cmdToggleTerm());